+++ autopush/websocket.py
@@ -37,6 +37,7 @@ from functools import wraps
 
 import cyclone.web
 from autobahn.twisted.websocket import WebSocketServerProtocol
+from boto.dynamodb2.exceptions import ProvisionedThroughputExceededException
 from twisted.internet import reactor
 from twisted.internet.defer import (
     Deferred,
@@ -522,9 +523,15 @@ class PushServerProtocol(WebSocketServerProtocol):
         d = self.deferToThread(self.ap_settings.router.register_user,
                                user_item)
         d.addCallback(self._check_other_nodes)
+        d.addErrback(self.trap_cancel)
+        d.addErrback(self.retry_hello)
         d.addErrback(self.err_hello)
         self.ps._register = d
         return d
+
+    def retry_hello(self, failure):
+        failure.trap(ProvisionedThroughputExceededException)
+        self.returnError("hello", "overloaded", 503)
 
     def err_hello(self, failure):
         """errBack for hello failures"""
@@ -631,7 +638,8 @@ class PushServerProtocol(WebSocketServerProtocol):
 
         # Are we paused, try again later
         if self.paused:
-            self.deferToLater(1, self.process_notifications)
+            d = self.deferToLater(1, self.process_notifications)
+            d.addErrback(self.trap_cancel)
             return
 
         # Process notifications differently based on webpush style or not
@@ -660,7 +668,8 @@ class PushServerProtocol(WebSocketServerProtocol):
         # Were we told to check notifications again?
         if self.ps._check_notifications:
             self.ps._check_notifications = False
-            self.deferToLater(1, self.process_notifications)
+            d = self.deferToLater(1, self.process_notifications)
+            d.addErrback(self.trap_cancel)
 
     def finish_webpush_notifications(self, notifs):
         """webpush notification processor"""
@@ -669,7 +678,8 @@ class PushServerProtocol(WebSocketServerProtocol):
             self.ps._more_notifications = False
             if self.ps._check_notifications:
                 self.ps._check_notifications = False
-                self.deferToLater(1, self.process_notifications)
+                d = self.deferToLater(1, self.process_notifications)
+                d.addErrback(self.trap_cancel)
             return
 
         # Send out all the notifications
@@ -740,6 +750,7 @@ class PushServerProtocol(WebSocketServerProtocol):
         d = self.deferToThread(self.ap_settings.make_endpoint, self.ps.uaid,
                                chid)
         d.addCallback(self.finish_register, chid)
+        d.addErrback(self.trap_cancel)
         d.addErrback(self.error_register)
         return d
 
@@ -755,6 +766,7 @@ class PushServerProtocol(WebSocketServerProtocol):
             d = self.deferToThread(self.ap_settings.message.register_channel,
                                    self.ps.uaid, chid)
             d.addCallback(self.send_register_finish, endpoint, chid)
+            d.addErrback(self.trap_cancel)
             return d
         else:
             self.send_register_finish(None, endpoint, chid)
@@ -1012,7 +1024,7 @@ class NotificationHandler(cyclone.web.RequestHandler):
 
         """
         client = self.ap_settings.clients.get(uaid)
-        if client and client.connected_at == int(connectionTime):
+        if client and client.ps.connected_at == int(connectionTime):
             client.sendClose()
             return self.write("Terminated duplicate")