+++ CHANGELOG.rst
@@ -2,6 +2,17 @@
 Autopush Changelog
 ==================
 
+1.6.1 (2015-09-28)
+==================
+
+Bug Fixes
+---------
+
+* Use proper client connected_at in delete handler.
+* Trap provisioned exception error in websocket handler.
+* Clear the node entry for a client even if we know the node is dead.
+* Ignore Cancelled exceptions.
+
 1.6.0 (2015-09-14)
 ==================